arrow key moves and move_to_pointer in BaseSelector, move_to_pointer wonky for LinearRegionSelector
